chore: move popover shim into __tests__/shims/
Gives file-based test shims a dedicated directory so future ones have
an obvious home. Drops the redundant `-shim` suffix from the filename
since the directory name already says "shim".
